openhcl: move firmware VMGS store/read/delete helpers into hibernate module
Move the VMGS hibernation-state helpers next to the token helpers:
- StoreFirmwareOutcome, store_firmware, read_firmware, delete_firmware now live
  in hibernate.rs (with unit tests for the store/read/delete round-trip).
- worker.rs keeps the VTL0/loader wiring (restore_vtl0_firmware_from_vmgs and
  set_vtl0_uefi_rip), calling hibernate::read_firmware for the bytes.
